Move location services to their own folders
