+++ services/location/getFormattedAddressFormCoords/index.js
@@ -1,5 +1,6 @@
 import Geocoder from 'react-native-geocoder';
-import utils from '../../utils';
+
+import utils from '../../../utils';
 
 export default function getFormattedAddressFromCoords(lat, lng) {
   return new Promise((resolve, reject) => {
